Return empty list for hw sync if no data read yet
When starting a new scan, there may be no data on the first ReadOne call.
Rather than returning `None` which results in an error being logged, return
an empty list.  This is the expected return type for hardware synchronisation.

For software triggers, we don't expect to have no data, so return still
return None in that case.