Work around scaling issue with long integration times
Integrating for longer than about 2.62 seconds requires the
FPGA's fixed-point accumulator to use a pre-divider value
of 2 (or bigger, depending on accumulation time).  This prevents
an overflow in accumulator.  The EM software configures the FPGA
correctly, but a bug in the EM software (introduced in v1.3.5) fails
to apply the inverse scaling factor to the values it reads out of
the FPGA.  This means that the data read via SCPI needs
to be scaled by the correct power of 2 to undo this.

The Em2 module, which does the low-level comms with the EM device, now takes
this into account, and automatically corrects the data.  It assumes the
bug will be fixed in the EM software from version 2.1.

Controllers:
- `Albaem2CoTiCtrl`:  This uses the Em2 module for comms, so this fix is applied.
- `Albaem2OneDCtrl`:  This does not use the Em2 module and has not been fixed.
   It will just raise an error for times > 2.621 seconds, regardless of EM software
   version.