Improve error checking in ZMQ receiver
Do some sanity checks on the data to ensure we don't miss
anything, and report failures via exceptions.  These
will propagate to users of the Em2 module, such as the
Albaem2CoTiCtrl controller.

We now clear the message queue (`_reset_worker_state`), in
the `start` method, before launching the thread.  This avoids
a race condition were the old queue could still be used while
the new thread is starting (or about to start).