Increase ZMQ receiver high-water mark
This allows the ZMQ sender (Electrometer) to keep sending data
even if the receiver (sardana controller) falls a little behind.

The default limit is 1000 messages, which is reached very quickly
at messages rates of 500 to 1000 Hz.  Increase it to 1M to
give the controller lots of time to catch up.  The controller
(and Sardana) tends to be slow to start, but then picks up
speed and can catch up.